requirements.txt
asgiref==3.4.1
backcall==0.2.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.1
decorator==5.1.0
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.44.0
django-phonenumber-field==6.1.0
django-rest-allauth==1.1
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
factory-boy==3.2.1
Faker==11.3.0
flake8==3.9.2
httpie==2.6.0
idna==3.3
ipython==7.30.1
itypes==1.2.0
jedi==0.18.1
Jinja2==3.0.3
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mccabe==0.6.1
numpy==1.21.5
oauthlib==3.1.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.24
ptyprocess==0.7.0
pycodestyle==2.7.0
pycparser==2.21
pyflakes==2.3.1
Pygments==2.10.0
PyJWT==2.3.0
PySocks==1.7.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
text-unidecode==1.3
traitlets==5.1.1
uritemplate==4.1.1
urllib3==1.26.7
wcwidth==0.2.5
